+++ kilo/static/kilo/js/babel-src/loading.js
@@ -1,0 +1,16 @@
+export class Loading extends React.Component {
+    render() {
+      if (this.props.show) {
+        return (
+          <div>
+            <div className="spinner-grow text-secondary m-3" role="status"></div>
+            <div className="spinner-grow text-secondary m-3" role="status"></div>
+            <div className="spinner-grow text-secondary m-3" role="status"></div>
+            <span className="visually-hidden">Loading...</span>
+          </div>
+        );
+      } else {
+        return null;
+      }
+    }
+}
